feat: Migrate Cached to KMP and Remove Android Specific Code
This commit migrates the `Cached` class to the `kmpuiviews` module, making it available for multiplatform use. Additionally, it removes code related to `Cached` from the `UIViews` module that is no longer needed.

-   **`Cached` Migration:**
    -   Moved `Cached`, `GenericCache`, and `ExpirableLRUCache` from `ContextUtils.kt` in `UIViews` to `Cache.kt` in `kmpuiviews`.
    -   Updated `OtakuCustomListScreen`, `DetailsViewModel`, `OtakuUtils`, and `NotificationFragment` to use the new `Cached` from `kmpuiviews`.
    - Removed the old `Cached`, `GenericCache`, and `ExpirableLRUCache` from `ContextUtils`.
    - Removed the `LinkedHashMap` override in `kmpuiviews` to use the basic one.
-   - Added a `Clock` check to the `kmpuiviews` code to check for time.
- - **UIViews:**
+    - Added a `Clock` check to the `kmpuiviews` code to check for time.
+
+- **UIViews:**
    - Updated `LocalCurrentSource`, and `LocalSourcesRepository` to be setup in the `KmpLocalCompositionSetup` in `kmpuiviews`
    - Removed unnecessary imports.
